+++ tests/unit/vcsEvent.spec.ts
@@ -1,17 +1,25 @@
+import sinon, { type SinonSandbox } from 'sinon';
+import { expect } from 'chai';
 import VcsEvent from '../../src/vcsEvent.js';
 
 describe('VcsEvent', () => {
-  let sandbox;
-  /** @type {import("@vcmap/core").VcsEvent} */
-  let event;
+  let sandbox: SinonSandbox;
+  let event: VcsEvent<void | object>;
 
   before(() => {
     sandbox = sinon.createSandbox();
-    event = new VcsEvent();
   });
 
   after(() => {
     sandbox.restore();
+    event.destroy();
+  });
+
+  beforeEach(() => {
+    event = new VcsEvent();
+  });
+
+  afterEach(() => {
     event.destroy();
   });
 
@@ -39,7 +47,7 @@ describe('VcsEvent', () => {
   describe('removing an event listener', () => {
     it('should remove a listener via API', () => {
       const size = event.numberOfListeners;
-      const listener = () => {};
+      const listener = (): void => {};
       event.addEventListener(listener);
       event.removeEventListener(listener);
       expect(event.numberOfListeners).to.equal(size);
@@ -47,7 +55,7 @@ describe('VcsEvent', () => {
 
     it('should remove a listener via remove', () => {
       const size = event.numberOfListeners;
-      const listener = () => {};
+      const listener = (): void => {};
       const remover = event.addEventListener(listener);
       remover();
       expect(event.numberOfListeners).to.equal(size);
@@ -62,11 +70,27 @@ describe('VcsEvent', () => {
     });
   });
 
+  describe('removing an event listener in a raiseEvent call', () => {
+    it('should not call the second listener', () => {
+      let removeListener2 = (): void => {};
+      const listener = (): void => {
+        removeListener2();
+      };
+      const spy = sandbox.spy();
+      event.addEventListener(listener);
+      removeListener2 = event.addEventListener(spy);
+      expect(event.numberOfListeners).to.equal(2);
+      event.raiseEvent(undefined);
+      expect(event.numberOfListeners).to.equal(1);
+      expect(spy).to.not.have.been.called;
+    });
+  });
+
   describe('raising of events', () => {
     it('should raise a void event', () => {
       const spy = sandbox.spy();
       event.addEventListener(spy);
-      event.raiseEvent();
+      event.raiseEvent(undefined);
       expect(spy).to.have.been.calledWithExactly(undefined);
     });
 
@@ -82,7 +106,7 @@ describe('VcsEvent', () => {
   describe('awaiting listeners', () => {
     it('should await listeners', async () => {
       const spy = sandbox.spy();
-      const listener = () =>
+      const listener = (): Promise<void> =>
         new Promise((resolve) => {
           setTimeout(() => {
             spy();
@@ -93,5 +117,19 @@ describe('VcsEvent', () => {
       await event.awaitRaisedEvent();
       expect(spy).to.have.been.called;
     });
+  });
+
+  it('should not call the second listener, if the listener is removed in the trigger', async () => {
+    let removeListener2 = (): void => {};
+    const listener = (): void => {
+      removeListener2();
+    };
+    const spy = sandbox.spy();
+    event.addEventListener(listener);
+    removeListener2 = event.addEventListener(spy);
+    expect(event.numberOfListeners).to.equal(2);
+    await event.awaitRaisedEvent(undefined);
+    expect(event.numberOfListeners).to.equal(1);
+    expect(spy).to.not.have.been.called;
   });
 });